journal,payment,receipt and report service added

+++ MeroHisab/MeroHisab.Core/Dto/JournalDto.cs
@@ -6,7 +6,7 @@
 
         public string Remarks { get; set; }
 
-        public long VoucherNo { get; set; }
+        public int VoucherNo { get; set; }
 
         public List<JournalDetailDto> Details { get; set; }=new List<JournalDetailDto>();
 
@@ -14,7 +14,7 @@
 
     public class JournalDetailDto
     {
-        public long LedgerId { get; set; }
+        public int LedgerId { get; set; }
         public decimal DrAmount { get; set; }
         public decimal CrAmount { get; set; }
